+++ eq-author/src/components/AnswerContent/AnswerProperties/Answers/RadioProperties.js
@@ -1,6 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
-import Required from "components/AdditionalContent/Required";
+import Required from "components/AnswerContent/Required";
 import styled from "styled-components";
 
 const Container = styled.div`
