+++ eq-author/src/components/AnswerContent/AnswerProperties/Answers/index.js
@@ -2,4 +2,4 @@ export { default as Decimal } from "../Decimal";
 export { default as DateProperties } from "./DateProperties";
 export { default as UnitProperties } from "./UnitProperties";
 export { default as DurationProperties } from "./DurationProperties";
-export { default as TextProperties } from "./TextProperties";
+export { default as TextFieldProperties } from "./TextFieldProperties";
